requirements.txt
transformers==4.38.2
sentence-transformers==2.6.1
langchain==0.1.20
langchain-community==0.0.38
langchain-text-splitters==0.0.1
faiss-cpu==1.7.4
python-dotenv==1.0.1